requirements.txt
Django==1.11.10
djangorestframework==3.7.7
Markdown==2.6.11
pytz==2017.3
